moved mongo functions to mongo_utils.py and updated docs
